[Move] Extract the range-proof machinery into range_proof.move
`encrypted_amount.move` mixed two concerns: the encrypted-amount type ladder
and the Bulletproof batching that range-checks it. The second only ever touches
`Element<G>` commitments — it knows nothing about limbs — so it moves out:

  RangeProofs, new_range_proofs, batch_sizes, the Bulletproof verifier,
  BULLETPROOFS_VERSION, LIMB_BITS -> RANGE_BITS,
  MAX_RANGE_PROOF_BATCH_SIZE -> MAX_BATCH_SIZE, assume_range_checked

`verify_in_range` now collects the limb commitments and delegates to
`RangeProofs.verify`, which consumes the struct — so the `proofs` field is
private to the new module rather than destructured by its caller.

Note `new_range_proofs`' abort moves from `encrypted_amount::ERangeProofRequired`
(code 6) to `range_proof::ERangeProofRequired` (code 0). Nothing maps abort codes
today, on chain or in the SDK.

Regenerated `ts-sdk/src/contracts/`; `helpers.ts` now calls
`range_proof::new_range_proofs`. 39/39 Move tests pass.
