+++ ts-sdk/src/contracts/contra/encrypted_amount.ts
@@ -1,10 +1,9 @@
 /**************************************************************
  * THIS FILE IS GENERATED AND SHOULD NOT BE MANUALLY MODIFIED *
  **************************************************************/
-import { bcs } from '@mysten/sui/bcs';
 import { type Transaction, type TransactionArgument } from '@mysten/sui/transactions';
 
-import { MoveStruct, normalizeMoveArguments, type RawTransactionArgument } from '../utils/index.js';
+import { MoveStruct, normalizeMoveArguments } from '../utils/index.js';
 import * as twisted_elgamal from './twisted_elgamal.js';
 
 const $moduleName = '@local-pkg/contra::encrypted_amount';
@@ -38,35 +37,6 @@ export const VerifiedAmount = new MoveStruct({
 		pk: twisted_elgamal.PublicKey,
 	},
 });
-export const RangeProofs = new MoveStruct({
-	name: `${$moduleName}::RangeProofs`,
-	fields: {
-		proofs: bcs.vector(bcs.vector(bcs.u8())),
-	},
-});
-export interface NewRangeProofsArguments {
-	proofs: RawTransactionArgument<Array<Array<number>>>;
-}
-export interface NewRangeProofsOptions {
-	package?: string;
-	arguments: NewRangeProofsArguments | [proofs: RawTransactionArgument<Array<Array<number>>>];
-}
-/**
- * Wrap `proofs` into `RangeProofs`; rejects an empty set so the range check can't
- * be skipped on chain.
- */
-export function newRangeProofs(options: NewRangeProofsOptions) {
-	const packageAddress = options.package ?? '@local-pkg/contra';
-	const argumentsTypes = ['vector<vector<u8>>'] satisfies (string | null)[];
-	const parameterNames = ['proofs'];
-	return (tx: Transaction) =>
-		tx.moveCall({
-			package: packageAddress,
-			module: 'encrypted_amount',
-			function: 'new_range_proofs',
-			arguments: normalizeMoveArguments(options.arguments, argumentsTypes, parameterNames),
-		});
-}
 export interface NewEncryptedAmountArguments {
 	l0: TransactionArgument;
 	l1: TransactionArgument;
